Music Puzzle: On wrong note, match again at start of melody
Suppose that the sequence of notes that you have to play is C D E G, and you
play C D C D E G. This should be accepted as a valid solution, because although
the third note played is wrong for the 3rd note of the melody, it's the correct
first note. But this regressed in commit bf96e5f.

After failing to match the note in the middle of the melody, match it again at
the start.

Fixes https://github.com/endlessm/threadbare/issues/437