Stealth quest & template: Remove custom transition
In the Stealth of Lore quest there are 2 transitions happening when the
scene loads: one triggered by the collectible using SceneSwitcher, and a
custom one. The custom transition also has a %TransitionFade node as a
dependency in the script.

Remove custom fade-in transition from the Stealth script. One is enough.

Remove custom fade-out transition when the player is detected. Use the
new SceneSwitcher.reload_with_transition instead.

Remove the corresponding nodes from Lore and template Stealth scenes.

Note: This changes the transition from half a second to one second. We
can add a parameter all the way down to the tween, but it doesn't seem
worth for now.