Rename StoryQuestStarter to Elder
The previous name was based on my incorrect/obsolete understanding that there
would be one NPC per quest and that all quests were StoryQuests.

We now have the general concept of quests, which are divided into StoryQuests
and LoreQuests. The NPCs who send you on quests are elders.

Rename the directory, script and scene accordingly. Keep the names for the 2
different dialogues: currently it is correct that one dialogue is for a
LoreQuest and another for a StoryQuest.