Trees: Use animated sprite
Change the node in the tree scene from Sprite2D to AnimatedSprite2D. And
rename it because it had the same name as the root node "Tree".

Add a new setting in the wind shader to pass the amount of frames
because now the texture width needs to be divided by it. Otherwise the
shader deforms the tree too much.

Because of this change the tree wool variant is removed from the scenes
that had it applied, because it's a static PNG.