Tree: Export internal sprite frames
So it can be changed per tree.

Save current green sprite frames in a resource file, and reference it as
the default through its UID.

Also set the animation to 10 frames per second, which is how all
animations are meant to be.

Also call play(), when setting the sprite frames. This is needed when
the aprite frames change. It also
makes the frame-by-frame animation be played in the editor.