Bush: Use animated sprite
Like the trees. Remove the current Decoration script and add a new one
specific for the bush. To 1. expose the sprite frames and 2. to
randomize the initial frame.

Make it a tool so the animation plays in the editor. But also add a
pre-save handler to reset the frame progress before saving.

Save the sprite frames in a separate resource file, because alternative
files are coming.