Add new rock sprite to music puzzle
This commit adds the new rock sprites to music puzzle. Since they are animated, the sprite had to be converted into an animated sprite. The animation player was removed.

Since Rocks could be a decoration, their sprites were left in the assets folder, so they can easily be reused.

Also: Moved the musical rock xylophone to the "components" folder