Fix attack animation getting stuck when changing modes
Previously, if the player's AnimationPlayer was set to the blow animation and it changed mode from FIGHTING to COZY, it would stay looping in the blow animation.

Changed the code so it can never be in a fighting animation if the mode is COZY by matching by the player_mode first.

Then, changed the logic of the fighting mode so it handles the cases in which the player should walk because they are not attacking that frame.

Finally, adding a comment to explain how if the blow animation is played, it's only cut off if it is at the anticipation phase.

Fixes https://github.com/endlessm/threadbare/issues/421