Add storybook to select a StoryQuest to play
This is not actually wired up to the game, because there is currently only one
StoryQuest in the source tree. The intention is that, once we have more than
one, interacting with the StoryQuest elder would give you this interface (or an
improved version of it) to choose which one to play.

The purpose of adding this now is to define the "Quest" resource type and place
it in the StoryQuest template directory.

This could also be used to populate the Credits screen
(https://github.com/endlessm/threadbare/issues/402).

Fixes https://github.com/endlessm/threadbare/issues/470