Add AreaFiller component
This is an editor tool which uses Poisson disc sampling to generate
random but evenly-spaced points within a CollisionObject2D, and then
instantiates scenes sampled randomly from a list at those points.
Following the pattern we use throughout this project, it can
additionally choose from a list of SpriteFrames to assign to each
scene's sprite_frames property, if it has one.

The sampler is reasonably fast but, on my system, takes about 40ms to
fill in a reasonably-sized polygon – multiple frames, too long to do in
the main thread. I have put very little effort into optimising it, and
none into moving it off the main thread; instead I'm just declaring that
it should not be used at runtime.

I intend this to be used to generate the boundary forests seen in Fray's
End and the music puzzle, which were previously placed by hand; and to
generate the flowerbed in Fray's end, which was also previously placed
by hand.

It is a node which acts on a parent CollisionObject2D rather than being
a script that is attached to a CollisionObject2D because the forest
should be a StaticBody2D (so that the player can't walk through it –
that's the point) while the flowerbed can be an Area2D which has no
physics of its own.

https://github.com/endlessm/threadbare/issues/556