Teleporter: Rename "scene_to_go_to" to "next_scene"
This name is clearer, shorter, and is capitalised less erratically by the Godot
Inspector.

Strictly speaking it may not be "next" in the sense of "successor" --
teleporters could be used to split a single "level" into multiple
scenes, as we previously did with the world map -- but in practice we
only currently use Teleporter to connect levels linearly, and the
"next_scene" naming is consistent with other scene-switching components.

https://github.com/endlessm/threadbare/issues/483