Stealth logic: Remove global instant detection
This was setting every guard to instant player detection to true or
false since commit 67c81f46. But that prevents mixing guards with
instant detection in the same scene, making the modding inflexible. With
this change, the 2 guards next to the bridge in the Lore Stealth game
should detect the player instantly. Thus this fixes the regression.

Fix https://github.com/endlessm/threadbare/issues/502