Redesign LoreQuest music puzzle (#506)
Redesign music puzzle to simplify dialogue, choices, and solutions, and
allow the player to interact with the signs before they have been lit.
Redesign tilemap to reduce complexity.
Redesign aesthetics and add ruined buildings to feel more worn and
broken, as the narrative suggests.

These changes were made in response to feedback of complexity both in
tilemap design and dialogue, as well as a need to meet the aesthetics of
the narrative for this element of the game.

---------

Co-authored-by: Will Thompson <wjt@endlessos.org>