Ink combat: delay door opening to make it more noticeable
Reconnected goal_completed signal to an animation instead of directly to
the door, so we can delay the door to be opened after.
The rationale behind this is that too many things might be happening
when the challenge is completed: inkwell disappearing, enemies being
defeated, ink blobs still in screen. So, waiting a bit to open the door
makes it easier to notice.

Related to https://github.com/endlessm/threadbare/issues/76