MusicalRock: Remove automatic modulation by note
Previously, the colour of each MusicalRock was set in the rock's script,
which set the modulate property on the sprite. The colour was chosen by
picking 7 evenly-spaced hues for the 7 possible notes, and the same
saturation and value for each.

This was convenient, especially since we only have one rock asset. But
it makes it harder to have different appearances in different instances
of the musical-rock/sequence puzzle. For example, you may want the
objects to be four different shapes, not recoloured versions of the same
asset.

Remove the logic that recolours the rock based on its note. Instead,
modulate each rock in the scene that uses it, using the same colour that
was previously generated from the note in code.

This looks a bit weird in the editor because the rock's collision shapes
and "kick" interaction label also get modulated. But at runtime the
collision shapes are invisible, and the label is not actually used – it
just sets the position where the label belonging to the player will be
positioned. And I consider modulating the sprite to be a temporary
measure; it would be better to recolour the rocks.

(I tried using PropertyUtils to export the modulate property of the
AnimatedSprite2D. It solved this problem but in my opinion the cure was
worse than the disease: it added more complexity to what is otherwise a
very simple scene, and the resulting "Sprite Modulate" property in the
inspector had no documentation.)

Fixes https://github.com/endlessm/threadbare/issues/520