Music Puzzle: simplify demoing melody
Interacting with an altar causes the corresponding melody to be played
on the rocks. This should not be considered player input.

Previously, this was implemented by setting an _is_demo flag during the
demonstration and ignoring the note_played signal if it is set. But in
my opinion a clearer solution is to only emit the note_played signal
when the player actually interacts with the rock.

Move the signal emission from MusicalRock.play() to
MusicalRock._on_interaction_started(). Remove the _is_demo flag.