Music puzzle: Represent steps as nodes
Conceptually this music/sequence puzzle is a series of steps. Each step
has a sequence of nodes that the player must interact with in that
order, and a sign which provides a hint and lights up when that step is
solved. When all steps have been solved, the puzzle as a whole is
solved.

Previously, the sequences were represented as a string containing the
names of musical notes, which corresponded to a note property on
MusicalRock nodes. This makes the puzzle specific to musical melodies,
and is also error-prone: nothing prevented the level designer entering
the string "Sausages" as a "melody".

There was also a requirement that there be the same number of melodies
as bonfire signs, enforced with a warning in the editor.

Introduce a SequencePuzzleStep node, which contains the sequence for
this step and the corresponding bonfire sign. The sequence is
represented as an array of MusicalRock node references. This makes it
much more difficult to implement an unsolvable puzzle, and provides a
natural way to refer to the objects in the sequence directly. It also
makes it harder to get a mismatched number of signs and sequences. (But
not impossible: you could assign the same sign to multiple sequences.)

BonfireSign previously hardcoded a NodePath to its parent to find the
puzzle, which was convenient but error-prone – instead use a normal
reference, and disable interactivity if it is not provided.

Update both instances of the puzzle for these changes, and tweak the
template dialogue.

https://github.com/endlessm/threadbare/issues/522