Simplify learner templates (#515)
I reduced the amount of text and simplified the guidance within the
dialogue in the intro template, checkpoint, and combat template scenes
just to keep the "instruction" there to a bare minimum of bread crumbs
or hints. This is to go along with a previous PR where I adjusted the
language of the collectible dialogue for the stealth template scene as
well.

Within the combat scene and stealth scene I also generalized the player
name property to just be **Player Name** instead of Storyweaver and a
made a few nodes sound more generic to encourage more open
creativity/ideation for learners about the potential these components
have for their unique StoryQuests.

- Barrels are targets
- ThrowingEnemy is a ThrowingNPC
- I left "Guards" alone in stealth but will review with others on this
one as well.

---------

Co-authored-by: Will Thompson <wjt@endlessos.org>