MusicalRock: Allow customising sprite frames
This follows a similar pattern used elsewhere in the project, including
adding a warning if the expected animations are not defined.

This is primarily intended for StoryQuests but you could imagine we
could also have different-coloured versions of the base rock asset,
rather than using the modulate property as we currently do.

Fixes https://github.com/endlessm/threadbare/issues/524