Music puzzle: Only allow interacting with the next sign
Commit c4bdddc3f3da038902728762e834573fe8ccc891 changed the signs'
default behaviour so that each sign can be interacted with before it is
lit, allowing the player to see a demo of the corresponding melody. This
is a deliberate choice to make the puzzle easier, although it was made
configurable in case someone wants a harder puzzle.

However, that change introduced a problem. At the start of the puzzle,
only the melody for the first sign is accepted as input, but all four
signs were interactive, misleading the player.

Add an interactive_hints property to the MusicPuzzle logic node,
controlling whether the sign for the current step of the puzzle should
be interactive. Add code to propagate this through to the BonfireSigns.
Rename the corresponding property on BonfireSign and stop exporting it –
this is now a whole-puzzle setting, not a per-sign setting.

Fixes https://github.com/endlessm/threadbare/issues/523