MusicalRock: Wobble for a fixed duration
Previously, the play() method would resolve after:

- The rock's audio stream has played to completion; then
- The current loop of the struck animation has finished.

This meant that, when a sequence is being demoed, if the audio streams
for each rock are different lengths (which they are – quite markedly so)
then the demo does not play at a consistent speed. This looks & sounds
bad, and will also makes it hard to synchronise a hint animation on the
sign that the player has interacted with to play the demo.

Make the play() method (which is async) take a fixed length of time,
configured via a property which defaults to 1 second. Update the
wobble_silently() implementation (which was already fixed at 1 second)
to run for the same time, using the same logic.