Game elements: use templates as default
For the exported properties, use defaults as in the template quest. For
example, the player name now defaults to "Player Name" instead of
"StoryWeaver". Same for exported SpriteFrames in player and in enemies.

The next commit will adjust the scenes for both Lore quest and the
StoryQuest template.

Note: The exception is exported sounds. These will be addressed later
because they will be changed to "no sound" as default, so the scripts
need to consider that the audio streams can be null.