Music puzzle: Allow customising BonfireSign sprite
Previously, the bonfire sign was made up of three sprites:

- The plain, unadorned, unlit altar (a static image).
- The glyph on the front of the sign, which gives a hint as to that step
  of the puzzle; this was 4 sets of static images, for each hint in an
  unlit or lit state.
- The fire animation which appears when the corresponding puzzle step is
  solved.

To make it feasible to reuse this prop for other sequence puzzles, with
different sprites, a different approach is needed:

- Import all the previous elements into a single Aseprite file, with
  different layers for the 4 different hints.
- Export the result as 4 different sprite sheets.
- Simplify the BonfireSign scene to a single AnimatedSprite2D.
- Use the generic sign image created by Heather as a default.
- Export a SpriteFrames property from the BonfireSign.
- Set this property in the LoreQuest puzzle to keep its appearance and
  behaviour the same.

The new sign asset, and its SpriteFrames, have a "hint" animation. This
is not used in this commit but will be in a future commit.

Fixes https://github.com/endlessm/threadbare/issues/521