pre-commit: Enforce checking for large files
The default behaviour for check-added-large-files is to ignore files
that are already in the tree, even if `--all-files` was passed to
`pre-commit run` (as we do in CI). This caused us to add a 17 MB WAV
file to the source tree.

Add the necessary flag to make this hook work correctly.

Fixes https://github.com/endlessm/threadbare/issues/589