Swap COPYING and LICENCES/MPL-2.0.txt
Previously, LICENCES/MPL-2.0.txt contained the actual text of the MPL,
and COPYING at the top level was a symbolic link to it. The problem with
this is that the GitHub web UI does not follow symbolic links when it
tries to determine the license for the project, so the "License" tab
just contains the name of the symlink target.

Now it is the other way around: COPYING is the text of the license,
while LICENSES/MPL-2.0.txt (which is used only by the `reuse` tool) is a
symbolic link.