Rename MusicalRock to SequencePuzzleObject
This is no longer necessarily a rock, nor musical, but a generic object
that can be kicked and may or may not make a musical sound when the
player does so.

Rename the scene & class; adjust its API to reflect its new generic
nature. Rehouse it from a xylophone directory in the lorequest puzzle
into the components tree. Move the rock sprite frames up one level of
the hierarchy, to the lorequest puzzle's components directory.

https://github.com/endlessm/threadbare/issues/526