Adjust musician hint levels
Previously, the get_limited_hint_level() method provided by the GDScript
code would return the number of previous hints at the current step of
the puzzle, mod 3. This made it impossible to have a different number of
stages of hint in the dialogue.

There was also an issue in the current version of the dialogue where the
3rd hint for the 1st step was:

> Musician: How wonderful! We lit the first altar! Can you work out the
> second part of the melody?

but by definition if we were showing that line of dialogue, the first
altar was unlit.

Change the method to return the raw number of previous hints. Adjust the
dialogue tree to keep repeating the final hint for each step once all
previous hints have been exhausted. Move the incorrect hint mentioned
above to be the first of for hints for the second step, and make the
third hint for the first step a variant of the second.