HintSign: Change interact prompt to Examine
It is now possible that interacting with the sign does not make any
sound, for two reasons:

1. SequencePuzzleObjects do not need to have an AudioStream defined; or

2. The HintSign can be configured not to demonstrate the sequence.

So the "Listen" interact prompt is not always appropriate. It also
doesn't really make sense in the original context of a stone plinth: you
wouldn't exactly listen to a lump of stone!

Use a more generic "Examine" prompt instead.