hiding_mushroom: Simplify behaviour
Previously, the behaviour of the mushroom was configurable: it could be
configured to start hidden, or start visible; and it could be configured
to stay hidden/revealed once its state changes. I found the logic
confusing, particularly the dynamic redefinition of the "stay
hidden/revealed" property based on the value of is_hidden. And the
behaviour of the four instances in Fray's End was incorrect: they are
supposed to start revealed, and hide when the player gets near.

Remove all configuration from the mushroom. Make it always follow the
behaviour of being visible unless the player is nearby.

Fixes https://github.com/endlessm/threadbare/issues/580