Checkpoint: Use KnitWitch asset
Replace the Sprite2D with an AnimatedSprite2D. Export a sprite_frames
property on the top-level scene, following the pattern used in other
scenes. Set the default to the purple KnitWitch. Play the appear
animation from activate(), but only if the sprite was previously
invisible.

This preserves the previous behaviour that if you move backwards in a
level, past an earlier checkpoint, that checkpoint will be re-activated.
I think this is a little weird but it's not a huge problem.

Fixes https://github.com/endlessm/threadbare/issues/603