Checkpoint: Remove collision
I introduced this in commit 2dedba48be4249be5b62008e94cbb6ebc7a8c549
("Indicate checkpoints visually") but it has proven to be quite annoying.
Remove it, and allow the player to walk through the KnitWitch. It's a
little weird but since the KnitWitch asset is a floating witch we can
probably handwave it away with "the KnitWitch is visible but
incorporeal" or something.

Helps https://github.com/endlessm/threadbare/issues/455