Sokoban: Reduce vertical size of level 3a
This level was previously so tall that the bottom boundary was not
visible on the screen.

Remove one row from the centre of the puzzle. This does not
fundamentally change the puzzle, but it makes all walkable cells be
visible, and at least the bottom/top of the top/bottom boundary
(respectively).

Fixes https://github.com/endlessm/threadbare/issues/645