Title: Use loud version of main theme
The quiet version we were using before is mastered at a substantially
lower volume than the loud version, and (more to the point) the other
pieces of in-game music. This meant that an appropriate volume for the
main game is too quiet for the title screen; and that if you crank up
the music volume on the title screen to compensate, you are blasted by
the in-game music.