StoryQuest Template: Add template_ prefix to all files
In discussion with Heather and Justin, it is useful to consistently name
all files that should be renamed with a prefix that should be replaced
with the short name of the quest, e.g. replacing template_ with stella_,
to make it clearer that the correct file is being edited.

Rename all files and directories in the StoryQuest template to contain
template_ as a prefix.

The one exception is quest.tres which must not be renamed – that is how
the storybook discovers quests.