Player: Add DEFEATED mode
Add a new mode to the player. In this mode the player can't be
controlled anymore, and a one-off animation will be played.

Set velocity to zero and ignore any input in this new mode.

Add a new signal to player that is emited when the mode changes. And
connect it from the script for the player AnimationPlayer. This is to
avoid using the _process in the case of DEFEATED, which just needs to
call play() on the specific animation once.