Player: Define is_running based on current speed
Previously, is_running() was defined solely based on whether the user is
holding the "running" action (Shift or L1). However, in combination with
an analogue joystick, it is possible to move the player slowly while
holding L1, in which case the running animation looks weird.

Instead, define is_running() based on the player's current speed being
greater than the walk speed. This means that if the player holds L1 but
inclines the joystick only slightly, the normal walk animation plays and
no dust particles are emitted.