Collectible: Don't reset sprite texture every frame
Previously, collectible_item defined a _process() method which would set
the sprite_2d's texture to the appropriate texture for the InventoryItem
assigned to CollectibleItem.item.

This is unnecessary: the texture only needs to be updated when
CollectibleItem.item is set.

Move this logic into the _set_item setter. Call it from _ready(),
instead of _update_interact_action() (which it also calls), since the
sprite_2d (and interact_area) are only defined when the node is ready.

Fixes https://github.com/endlessm/threadbare/issues/295