Music puzzle: Move all altars 16px higher
In a playtesting session, a younger player struggled with the music
puzzle simply because she kept getting stuck between the altar and the
rocks and triggering the altars instead of the rocks. (I've run into the
same thing myself, though I'm rushing impatiently through the melodies
so I didn't pay much attention when it happened to me.)

Move all the altars 16px (1/4 tile) higher.