Music puzzle: Use house scene for house props
This scene has many homes in various states of ruin.
Beneath the stochastic whispers of the buildings' slow collapse, they
did not previously use the house prop scene; instead, each one was a
Sprite2D, with collisions handled by a separate StaticBody2D which had
one collision polygon behind each house. Those polygons traced the
outline of each house, making it impossible to walk behind them. In
addition, the static body was in the wrong collision layer, meaning each
house was considered a player for the sake of collisions.

Replace each one with an instance of a house scene, all inside a new
Houses node. The only gameplay change is that you can now walk behind
the houses; and for the level designer, it is easier to edit the level.