Music puzzle: Move thread to be visible from final note
Previously, in most positions where you can kick the left-most rock to
finish the final step of the puzzle, the collectible was off-camera when
it popped into view.

Move it across so that at the furthest-left and -up point where you can
interact with the leftmost rock, the thread is visible.

Fixes https://github.com/endlessm/threadbare/issues/578