Quest: Simplify first_scene property
Previously, first_scene could be an absolute path to a scene, or a path
relative to the quest.tres file. My idea was that StoryQuest creators
would typically keep the file structure of the template quest, so we
could expect the first scene to always be at `0_intro/intro.tscn`
relative to the quest.tres file.

But commit 3630e5141218615b48c082676a34fc53c2bb092a changed this. The
decision was made that all files in the template should have a
`template_` prefix, and that after duplicating the template, StoryQuest
creators should rename every file in their copy, replacing `template_`
with something specific to their quest. It follows that we can't know
ahead of time what the intro scene's relative path will be.

Remove this logic & the default value for first_scene. Set the property
in the template quest.

Since this is now a property that StoryQuest creators will have to care
about, move the property out of the collapsed-by-default Advanced group.