Quest: Add sprite_frames & animation_name properties
This will be used to show an animation for the quest in the elder's
book of stories. The sprite_frames is optional, and defaults to null.

Make quest.gd a @tool script, so that we can offer a drop-down list of
animation names in the inspector based on the animations defined by the
SpriteFrames.

https://github.com/endlessm/threadbare/issues/557