Storybook: Fix loading quests in exported project
Previously, I passed `"Quest"` as the optional type_hint parameter to
ResourceLoader.exists() and .load(). When running in the editor, this
works, but in the exported project it does not work for the LoreQuest's
quest resource. If I remove the parameter from exists() but leave it on load(), I see:

    ERROR: No loader found for resource: res://.godot/exported/133200997/export-9a44ecd99f0d01072e980456714013f7-quest.res (expected type: Quest)
       at: _load (core/io/resource_loader.cpp:359)
    ERROR: Error loading resource: 'res://scenes/quests/lore_quests/quest_001/quest.tres'.
       at: load (core/core_bind.cpp:76)

I think the problem here is that "Quest" is not a distinct resource type
known to the engine. The behaviour is different for the StoryQuest template
(res://scenes/quests/story_quests/template/quest.tres) because this
resource is preload()ed by UID in this file, so the engine returns the
cached version.

Remove the hint.