Stella: Move InteractAreas for tortoise segments
The 7 segments of the tortoise are much bigger than they appear to be,
and are all positioned at the same place in game space. As a result, the
7 InteractAreas were also at the same position in game space. This means
that the logic in InteractZone which chooses the "best" InteractArea
when more than one is overlapping doesn't work: it chooses the closest
one, but if they're all at the same position there is no single closest.

Move the InteractAreas to be in the position of their visible segment,
and set the child collision shapes' relative positions to (0, 0). This
keeps the interact areas the same (though I did tweak a couple of them),
while fixing the logic that detects the closest.

It also makes it simple to position the "Press" prompts by setting them
all to be right next to the InteractArea.

https://github.com/endlessm/threadbare/issues/681