Stella: Skip opening Stealth dialogue when spawning at checkpoint
This scene uses a Cinematic to play an opening narration. Cinematic
plays its dialogue unconditionally. But it is weird to have the
narration be replayed after you are caught and respawn.

We can't always distinguish between starting the level for the first
time and reloading it, but if the player has reached a spawn point we
can detect that from the dialogue since the GameState global is
accessible there. Skip the narration in that case.