ci: Disable linting & REUSE checks in forks
We cannot expect learners to have set up pre-commit in their checkouts
to fix errors before they push a branch. But then having a build show as
failed due to (typically) trailing whitespace in a dialogue file is not
very helpful.

Similarly, the REUSE linter failing in forks is not especially helpful.
In fact it might be actively harmful because people may be tempted to
put any old licensing information in that will silence the linter. We
will need to apply human review to what's being contributed upstream
regardless of what `reuse lint` says.

We still want to enforce both of these checks for changes that go
upstream - but when a change is submitted upstream, we can apply linter
fixes and confirm licensing information ourselves.

Fixes https://github.com/endlessm/threadbare/issues/590