Add a single rock to the intro scene template (#697)
Add a single rock to the intro template as an exemplar of a simple
background or world "component". This rock.tscn can be a reference for
how to create a simple art asset using:

StaticBody2d
Sprite2D
CollisionShape2D

This type of scene may be an approachable "first scene from scratch"
example.