Pause overlay: Add button to go back to title screen
And move the Resume button to the top, as it is the default focused
option.

Currently this is a way to abandon and lose all your progress, because
nothing is persisted. But is also a way to restart the experience if you
get stuck in the Sokobans, for example.

In the future, the progress should be persisted, and the title screen
should be able to offer a Resume + Restart options instead of Start, if
there is progress.

Fixes https://github.com/endlessm/threadbare/issues/432