GameState: Remove second instance of Inventory
There were 2 inventories of items, one global and one for the current
quest. And collected items were added to the two of them, while the
per-quest one was cleared when a quest starts. But in practice, we are
currently having a single inventory.

So, leave a single instance of the inventory in the game state. And
clear it when a quest starts.