GameState: Persist amount of items in inventory
At this point, we only need to count how many items are in the
inventory to persist it. Assume that each quest provides 3 items all
of different types (Memory, Imagination, Spirit) for the player to
collect them.

The maximum slots in the HUD is hardcoded to 3 as well.

The items are added to the inventory in the order they are collected.

However, since we are saving only the amount of items collected, the
order when restored will always be the same (first Memory, etc). So
allow adding an item of same type more than once. Also the next commit
will change the requirement to add the items to the Eternal Loom by
amount.