SceneSwitcher: Restore game state from web hash
If the scene in the URL hash matches the scene to restore, then act as
if the player continued the game from the main menu.

Otherwise, assume that the player is testing or debugging an isolated
scene from a given URL. This is the same behavior as if a scene that is
not the main one is ran from the editor.