Guard: Improve animations
Unfortunately, the sprite frames for Storyvore and the guard template
have animations of different length:
- Attack: 6 and 5 frames
- Idle: 11 and 8 frames

The only animation with the same length is the walking animation (6
frames). This is because the template was based on a placeholder, not in
the final character.

So, match the frame by frame animations in AnimationPlayer to the longer
animations (the 2 from Storyvore). The template will have the last frame
still for a few frames.

Add each frame to the AnimationPlayer timeline. Changing the animation
and adding 2 frames makes the animation play but not at a consistent
10 FPS.

This fixes a bug where the idle animation wasn't playing, because it was
done by calling play in the SpriteFrames node directly, and at the same
time the AnimationPlayer was calling stop(), which made the character be
still in frame zero of the idle animation.

Also: Remove waiting for 0.4 seconds before playing the attack
animation. Otherwise it looks odd that the player defeated animation
happens before it.