Warn if the project has been cloned without Git LFS
A common problem is to clone the project on a system that does not have
Git LFS installed, and then be confused by error messages reading "Load
failed due to missing dependencies".

Add an editor plugin that detects this case, by checking whether
icon.png at the root of the repository (accessed by UID, following our
convention) contains a Git LFS pointer. In this case, show a pop-up
window with a warning. When the game developer clicks Learn More… in the
dialogue, take them to an appropriate thread on our discussion forum.

https://github.com/endlessm/threadbare/issues/231