Teleporter: Add documentation for spawn_point_path
Previously, the `spawn_point_path` property was made accessible in the
inspector via _get_property_list, rather than `@export`. A downside of
doing it this way is that we cannot add documentation to the property
that shows up as a tooltip & in the documentation browser.

Instead, `@export` the property and override its attributes in
`_validate_property()` (so that it continues to be a drop-down of spawn
points in the target scene). Add a docstring.