ci: Attach Flatpak bundle to release
Since the Flatpak build was moved to the main export flow, it will be
run when a new release is made. (Previously it was intended to run in
this case, but did not for reasons outlined in that commit message.) But
the resulting `.flatpak` bundle was not added to the release as an
asset.

Do so!