ci: Attach web build to releases
Previously, for each new release we would build the game, but only add
the .pck file as a release asset.

Add the web build, as a .zip file, as an asset on each new release. This
will allow us (with some work on amalgamate-pages) to have the top-level
of the GitHub Pages site be the latest release.

Note that build artifacts are always zip files containing 1 or more
files; but when you use `actions/download-artifact` the zip file is
helpfully unpacked. Pack it back up with the same layout so it can be
used as input for amalgamate-pages.

Fixes https://github.com/endlessm/threadbare/issues/865