ci: Reduce wget output
By default, when stdout is not a TTY, wget uses the --progress=dot
style. It prints one dot for each 1K downloaded, with a total of 50 dots
per line.

Godot is tens of megabytes and the export templates are hundreds. That's
a lot of dots and it makes reading the action output when there happens
to be a cache miss here harder.

wget allows you to control how much data each dot and line represents.
In the "giga" style:

> each dot represents 1M retrieved, there are eight dots in a cluster,
> and 32 dots on each line (so each line contains 32M)

Use the giga style. This reduces the progress output for Godot itself to
2 lines:

    Saving to: ‘Godot_v4.4.1-stable_linux.x86_64.zip’
         0K ........ ........ ........ ........ 53%  330M 0s
     32768K ........ ........ ........ ...     100%  369M=0.2s
    2025-06-23 17:27:49 (346 MB/s) - ‘Godot_v4.4.1-stable_linux.x86_64.zip’ saved [62203199/62203199]

and the export templates to ~35 lines.